requirements.txt
# Core scraping
selenium==4.18.1
seleniumwire==5.1.0
webdriver-manager==4.0.1
playwright==1.42.0
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
fake-useragent==1.5.1

# Data processing
pandas==2.2.1
openpyxl==3.1.2
python-dotenv==1.0.1
pydantic==2.6.3
tqdm==4.66.2

# AI / NLP for enrichment & outreach
openai==1.14.3
anthropic==0.20.0
nltk==3.8.1
spacy==3.7.4

# Google APIs
google-api-python-client==2.122.0
googlemaps==4.10.0
google-search-results==2.4.2

# LinkedIn scraping helpers
linkedin-api==2.1.1

# Email validation & enrichment
pyIsEmail==1.4.1
hunter-python==0.3.0

# Proxy / anti-detect
requests-html==0.10.0
cloudscraper==1.2.71

# Database & caching
sqlite-utils==3.36
redis==5.0.3

# Scheduling & orchestration
schedule==1.2.1
celery==5.3.6

# Reporting & export
jinja2==3.1.3
weasyprint==62.1
matplotlib==3.8.3
seaborn==0.13.2
tabulate==0.9.0
rich==13.7.1

# HTTP retries
tenacity==8.2.3
backoff==2.2.1
